refactor(nvim): split mini config into ui and keybind modules
